Actually making AudioHelpers work

+++ Content.Shared/Audio/AudioHelpers.cs
@@ -1,6 +1,9 @@
 using Content.Shared.GameObjects.Components.Sound;
 using Robust.Shared.Audio;
 using Robust.Shared.GameObjects;
+using Robust.Shared.Interfaces.Random;
+using Robust.Shared.IoC;
+using Robust.Shared.Random;
 
 namespace Content.Shared.Audio
 {
@@ -11,7 +14,7 @@ namespace Content.Shared.Audio
         public static AudioParams WithVariation(float amplitude)
         {
             var scale = amplitude * IoCManager.Resolve<IRobustRandom>().NextGaussian();
-            return AudioParams.Default.WithPitchScale(scale);
+            return AudioParams.Default.WithPitchScale((float)scale);
         }
     }
 }
